Moved Classes from App.js to Their Own Modules
