+++ static/js/Board/BoardState.js
@@ -1,0 +1,10 @@
+export class BoardState {
+    constructor() {
+        this.moveEndPoint = false
+        this.drawEnabled = true
+        this.drawType = 'freeDraw'
+        this.inputEnabled = true
+        this.animating = false
+        this.animSpeed = 2
+    }
+}
