+++ static/js/app.js
@@ -1,387 +1,6 @@
-class Cell {
-    constructor(x, y, element) {
-        this.value
-        this.x = x
-        this.y = y
-        this.element = element
-        this.prev_value
-    }
-    update(value) {
-        if (!(value == 'wall-node' && this.isEndPoint())) {
-            this.prev_value = this.value
-            this.value = value
-            this.element.classList.add(value)
-        }
-    }
-    remove(value) {
-        if (value) {
-            this.value = this.prev_value
-            this.prev_value = null
-            this.element.classList.remove(value)
-        }
-    }
-    isEndPoint() {
-        return this.value == 'start-node' || this.value == 'end-node'
-    }
-    isWall() {
-        return this.value == 'wall-node'
-    }
+import { Board } from "./Board/Board.js"
+import { Button } from './Tools/Button.js'
 
-}
-
-
-class BoardState {
-    constructor() {
-        this.moveEndPoint = false
-        this.drawEnabled = true
-        this.drawType = 'freeDraw'
-        this.inputEnabled = true
-        this.animating = false
-        this.animSpeed = 2
-    }
-}
-
-
-class BoardMoveInput {
-    constructor(board) {
-        this.board = board
-        this.state = board.state
-        this.drawHandle = board.drawHandle
-        this.mousedown = false
-        this.elementHold = { cell: null, prev: null }
-
-        this.setMouseInput()
-    }
-
-    setMouseInput() {
-        this.mouseOverListener = this.onMouseOver.bind(this)
-        this.mouseDownListener = this.onMouseDown.bind(this)
-        this.mouseUpListener = this.onMouseUp.bind(this)
-        this.mouseLeaveListener = this.onMouseLeave.bind(this)
-        this.leaveContainerListener = this.onContainerLeave.bind(this)
-        console.log(this.board.cells)
-        for (let i = 0; i < this.board.cells.length; i++) {
-            for (let j = 0; j < this.board.cells[i].length; j++) {
-                const el = this.board.cells[i][j].element;
-
-                el.addEventListener('mouseover', this.mouseOverListener)
-                el.addEventListener('mouseleave', this.mouseLeaveListener)
-                el.addEventListener('mousedown', this.mouseDownListener)
-                el.addEventListener('mouseup', this.mouseUpListener)
-
-            }
-
-        }
-        this.board.element.addEventListener('mouseleave', this.leaveContainerListener)
-    }
-
-    mouseEvent(event) {
-        let cell
-        if (event.target.classList.contains('box')) {
-            cell = this.board.getCell(event.target.id)
-        }
-
-        if (!this.state.animating && this.mousedown) {
-            if (this.state.moveEndPoint) {
-                //console.log(this.elementHold.cell)
-                this.board.move(this.elementHold.cell, cell)
-            }
-            else if (cell.isEndPoint()) {
-                this.elementHold.cell = cell
-                this.elementHold.prev = cell.element
-                this.state.moveEndPoint = true
-            }
-
-            else {
-
-                if (this.state.drawEnabled && !cell.isEndPoint() && !this.state.moveEndPoint
-                    && !cell.element.classList.contains('start-node')) {
-                    switch (this.state.drawType) {
-                        case 'freeDraw':
-                            console.log('drawing wall')
-                            this.drawHandle.freeDraw(event, cell)
-                            break
-                        case 'erase':
-                            this.drawHandle.erase(event, cell)
-                            break
-                        default:
-                            console.log('No Draw Mode')
-                            break
-                    }
-
-                }
-            }
-
-        }
-    }
-
-    onMouseDown(event) {
-
-        this.mousedown = true
-        const x = this.board.getCell(event.target.id)
-        console.log(x)
-        this.mouseEvent(event)
-    }
-
-    onMouseUp(event) {
-        this.mousedown = false
-        this.state.moveEndPoint = false
-    }
-
-    onMouseOver(event) {
-        if (this.mousedown) {
-            this.mouseEvent(event)
-        }
-    }
-
-    onMouseLeave(event) {
-        if (this.state.moveEndPoint) {
-            this.elementHold.cell = this.board.getCell(event.target.id)
-            this.elementHold.prev = event.target
-            //event.target.classList.remove(this.elementHold.cell.value)
-        }
-
-    }
-
-    onContainerLeave(event) {
-        this.mousedown = false
-        this.state.moveEndPoint = false
-        // if (this.elementHold.cell) {
-        //     this.elementHold.prev.classList.add(this.elementHold.cell.value)
-        // }
-    }
-}
-
-
-class Board {
-    constructor(height, width) {
-        this.state = new BoardState()
-        this.cells = []
-        this.element = document.getElementById('grid');
-        for (var i = 0; i < height; i++) {
-            var row = document.createElement('div');
-            row.className = "row-container row";
-            row.id = "row" + i;
-            this.cells.push([])
-            for (var j = 0; j < width; j++) {
-                var box = document.createElement('div');
-                box.className = 'box box-node-container';
-                box.id = i + ',' + j;
-                row.appendChild(box);
-                const cell = new Cell(i, j, box)
-                this.cells[i].push(cell)
-            }
-            this.element.appendChild(row);
-        }
-        this.initializeBoard()
-        this.initalizeTools()
-        this.drawHandle = new DrawHandler(this)
-        this.mouseInput = new BoardMoveInput(this)
-
-    }
-
-    initializeBoard() {
-        // let walls = document.querySelector('.wall-node')
-        // walls.walls.forEach(element => {
-        //     element.classList.remove('wall-node')
-        // });
-        this.start = this.cells[11][12]
-        this.start.update('start-node')
-
-        this.end = this.cells[11][37]
-        this.end.update('end-node')
-    }
-
-    initalizeTools() {
-        this.addbtn = document.getElementById('add-wall')
-        this.removebtn = document.getElementById('remove-wall')
-        this.addWallListener = this.draw.bind(this)
-        this.removeWallListener = this.erase.bind(this)
-        this.addbtn.addEventListener('click', this.addWallListener)
-        this.removebtn.addEventListener('click', this.removeWallListener)
-
-        this.clearWallsbtn = document.getElementById('clearWalls')
-        this.clearVisitbtn = document.getElementById('clearVisit')
-        this.resetBoardBtn = document.getElementById('resetBoard')
-
-        this.clearWallsListener = this.clearWalls.bind(this)
-        this.clearVisitListener = this.clearVisited.bind(this)
-        this.resetBoardListener = this.resetBoard.bind(this)
-
-        this.clearWallsbtn = new Button('clearTool', this.clearWallsbtn, this.clearWallsListener)
-        this.clearVisitbtn = new Button('clearTool', this.clearVisitbtn, this.clearVisitListener)
-        this.resetBoardBtn = new Button('clearTool', this.resetBoardBtn, this.resetBoardListener)
-    }
-
-    getCell(id) {
-        let loc = id.split(',')
-        return this.cells[loc[0]][loc[1]]
-    }
-
-    move(fromNode, toNode) {
-        if (toNode) {
-            toNode.update(fromNode.value)
-            fromNode.remove(toNode.value)
-            switch (toNode.value) {
-                case 'start-node':
-                    this.start = toNode
-                    break
-                case 'end-node':
-            }       this.end = toNode
-        }
-
-
-    }
-
-    draw() {
-        if (this.state.drawType != "freeDraw") {
-            this.addbtn.classList.add('active-tool')
-            this.removebtn.classList.remove('active-tool')
-            this.state.drawType = "freeDraw"
-        }
-    }
-
-    erase() {
-        if (this.state.drawType != "erase") {
-            this.addbtn.classList.remove('active-tool')
-            this.removebtn.classList.add('active-tool')
-            this.state.drawType = "erase"
-        }
-    }
-
-    clearWalls() {
-        const walls = this.element.querySelectorAll('.wall-node')
-        walls.forEach(node => {
-            node.classList.remove('wall-node')
-        });
-    }
-
-    clearVisited() {
-        const visited = this.element.querySelectorAll('.visited-node')
-        //paths = this.element.querySelectorAll('.path-node')
-
-        visited.forEach(node => {
-            node.classList.remove('visited-node')
-            node.classList.remove('path-node')
-        });
-        // paths.forEach(node => {
-
-        // });
-
-
-    }
-
-    clearAll() {
-        const all = this.element.querySelectorAll('.box')
-        all.forEach(node => {
-            node.classList.remove('path-node')
-            node.classList.remove('visited-node')
-            node.classList.remove('wall-node')
-            node.classList.remove('start-node')
-            node.classList.remove('end-node')
-        });
-
-    }
-
-    resetBoard() {
-        this.clearAll()
-        this.initializeBoard()
-    }
-
-    animate() {
-        const tools = document.querySelectorAll('.tools')
-        console.log(tools)
-        tools.forEach(tool => {
-            tool.classList.add('disabled')
-        });
-    }
-
-    animateStop() {
-        const tools = document.querySelectorAll('.tools')
-        tools.forEach(tool => {
-            tool.classList.remove('disabled')
-        });
-    }
-}
-
-
-class DrawHandler {
-    constructor(board) {
-        this.board = board
-        this.tempWallList = []
-        this.initalPosition
-    }
-
-    freeDraw(event, cell) {
-        event.target.classList.add('wall-node')
-        cell.value = 'wall-node'
-    }
-
-    erase(event, cell) {
-        event.target.classList.remove('wall-node')
-    }
-
-    Point(x, y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    line(start, end) {
-        var p0 = new this.Point(Number(start.split(',')[1]), Number(start.split(',')[0]))
-        var p1 = new this.Point(Number(end.split(',')[1]), Number(end.split(',')[0]))
-        this.line(p0, p1)
-    }
-
-    // from https://www.redblobgames.com/grids/line-drawing.html
-    lineHelper(p0, p1) {
-        let points = [];
-        let N = this.diagonal_distance(p0, p1);
-        for (let step = 0; step <= N; step++) {
-            let t = N === 0 ? 0.0 : step / N;
-            points.push(this.round_point(this.lerp_point(p0, p1, t)));
-        }
-        return points;
-    }
-
-    diagonal_distance(p0, p1) {
-        let dx = p1.x - p0.x, dy = p1.y - p0.y;
-        return Math.max(Math.abs(dx), Math.abs(dy));
-    }
-
-    round_point(p) {
-        return new this.Point(Math.round(p.x), Math.round(p.y));
-    }
-
-    lerp_point(p0, p1, t) {
-        return new this.Point(this.lerp(p0.x, p1.x, t),
-            this.lerp(p0.y, p1.y, t));
-    }
-
-    lerp(start, end, t) {
-        return start + t * (end - start);
-    }
-    render(event, mousedown) {
-        // for (let i = 0; i < wallList.length; i++) {
-        //     const wall = wallList[i];
-        //     wall.classList.add('temp-wall-node')
-        // }
-    }
-}
-
-
-class Button {
-    constructor(type, element, f) {
-        this.type = type
-        this.element = element
-        this.element.addEventListener('click', f)
-    }
-    disable() {
-        this.element.classList.add('disabled')
-    }
-    enable() {
-        this.element.classList.remove('disabled')
-    }
-}
 
 
 class PathfinderInterface {
