+++ templates/base.html
@@ -48,7 +48,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
         
     <!-- <script src="{%static 'js/custom.js'%}" ></script> -->
-    <script src="{%static 'js/app.js'%}" ></script>
+    <script src="{%static 'js/app.js'%}" type="module"></script>
 </body>
 
 </html>
